factor out shared code across boards
