+++ platforms/allwinner-d1/core/src/drivers/mod.rs
@@ -1,2 +1,3 @@
+pub mod sharp_display;
 pub mod spim;
 pub mod uart;
